+++ ESPLibrary2.0/ESPLib/src/main/java/com/esplibrary/packets/PacketFactory.java
@@ -6,12 +6,15 @@
 package com.esplibrary.packets;
 
 import com.esplibrary.constants.PacketId;
+import com.esplibrary.packets.request.RequestAbortAudioDelay;
 import com.esplibrary.packets.request.RequestAllSweepDefinitions;
+import com.esplibrary.packets.request.RequestAllVolume;
 import com.esplibrary.packets.request.RequestBatteryVoltage;
 import com.esplibrary.packets.request.RequestChangeMode;
 import com.esplibrary.packets.request.RequestCurrentVolume;
 import com.esplibrary.packets.request.RequestDefaultSweepDefinitions;
 import com.esplibrary.packets.request.RequestDefaultSweeps;
+import com.esplibrary.packets.request.RequestDisplayCurrentVolume;
 import com.esplibrary.packets.request.RequestFactoryDefault;
 import com.esplibrary.packets.request.RequestMaxSweepIndex;
 import com.esplibrary.packets.request.RequestMuteOff;
@@ -31,9 +34,8 @@ import com.esplibrary.packets.request.RequestVersion;
 import com.esplibrary.packets.request.RequestWriteSweepDefinition;
 import com.esplibrary.packets.request.RequestWriteUserBytes;
 import com.esplibrary.packets.request.RequestWriteVolume;
-import com.esplibrary.packets.request.RequestAbortAudioDelay;
-import com.esplibrary.packets.request.RequestDisplayCurrentVolume;
 import com.esplibrary.packets.response.ResponseAlertData;
+import com.esplibrary.packets.response.ResponseAllVolume;
 import com.esplibrary.packets.response.ResponseBatteryVoltage;
 import com.esplibrary.packets.response.ResponseCurrentVolume;
 import com.esplibrary.packets.response.ResponseDataError;
@@ -57,7 +59,7 @@ import com.esplibrary.packets.response.ResponseVersion;
 public class PacketFactory {
 
     /**
-     * Creates a type ESP based on the specified packet ID and packet leghth.
+     * Creates a type ESP based on the specified packet ID and packet length.
      *
      * @param packetId      A PacketIds that represents a ESPPacket.
      * @return              An ESPPacket that corresponds to the packetId constant.
@@ -124,6 +126,10 @@ public class PacketFactory {
                 return new RequestAbortAudioDelay(packetLength);
             case PacketId.REQDISPLAYCURRENTVOLUME:
                 return new RequestDisplayCurrentVolume(packetLength);
+            case PacketId.REQALLVOLUME:
+                return new RequestAllVolume(packetLength);
+            case PacketId.RESPALLVOLUME:
+                return new ResponseAllVolume(packetLength);
             case PacketId.REQSTARTALERTDATA:
                 return new RequestStartAlertData(packetLength);
             case PacketId.REQSTOPALERTDATA: